feat: navbar hides automatically 🌂
Configure the navbar to be sticky _but_ to fade away after the user
scrolls out of the top of the page. The bar will return when the user
scrolls upwards or returns to the top of the page.

Additionally, when the mobile menu is opened, the fading behavior
is disabled.

This setup presently has a bug where opening the mobile menu
and then resizing the screen away from mobile results in the navbar
being persistent until the page is refreshed or the screen is
re-re-sized so the mobile toggle appears and then is closed.